Integrate

+++ implementationguide/src/main/java/com/google/android/fhir/implementationguide/IgManager.kt
@@ -24,6 +24,8 @@ import com.google.android.fhir.implementationguide.db.impl.entities.ResourceMeta
 import com.google.android.fhir.implementationguide.db.impl.entities.toEntity
 import java.io.File
 import java.io.FileInputStream
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import org.hl7.fhir.instance.model.api.IBaseResource
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.MetadataResource
@@ -32,14 +34,11 @@ import org.hl7.fhir.r4.model.ResourceType
 import timber.log.Timber
 
 /** Responsible for downloading, accessing and deleting Implementation Guides. */
-class IgManager internal constructor(igDatabase: ImplementationGuideDatabase) {
-
-  constructor(context: Context) : this(Room.databaseBuilder(context,
-                                                            ImplementationGuideDatabase::class.java,
-                                                            "implementationguide.db").build())
+class IgManager internal constructor(private val igDatabase: ImplementationGuideDatabase) {
 
   private val igDao = igDatabase.implementationGuideDao()
   private val jsonParser = FhirContext.forR4().newJsonParser()
+  private var defaultIgId: Long = -1L
 
   /**
    * * Checks if the [igDependencies] are present in DB. If necessary, downloads the dependencies
@@ -55,21 +54,19 @@ class IgManager internal constructor(igDatabase: ImplementationGuideDatabase) {
    */
   suspend fun install(igDependency: IgDependency, rootDirectory: File) {
     val igId = igDao.insert(igDependency.toEntity(rootDirectory))
-    rootDirectory.listFiles()?.forEach { file ->
-      try {
-        when (val resource = jsonParser.parseResource(FileInputStream(file))) {
-          is Bundle -> importBundle(igId, resource, file)
-          is Resource -> importResource(igId, resource, file)
-        }
-      } catch (exception: Exception) {
-        Timber.d(exception, "Unable to import file: %file")
-      }
-    }
+    rootDirectory.listFiles()?.forEach { file -> importFile(igId, file) }
   }
 
   /** Imports the IG from the provided [file] to the default dependency. */
   suspend fun install(file: File) {
-    TODO("not implemented yet")
+    // lazy suspend initialization
+    defaultIgId = igDao.getImplementationGuide(DEFAULT_DEPENDENCY.packageId,
+                                               DEFAULT_DEPENDENCY.version)?.implementationGuideId
+      ?: -1
+    if (defaultIgId == -1L) {
+      defaultIgId = igDao.insert(DEFAULT_DEPENDENCY.toEntity(File("NotARealDirectory")))
+    }
+    importFile(defaultIgId, file)
   }
 
   /** Loads resources from IGs listed in dependencies. */
@@ -95,10 +92,27 @@ class IgManager internal constructor(igDatabase: ImplementationGuideDatabase) {
   suspend fun delete(vararg igDependencies: IgDependency) {
     igDependencies.forEach { igDependency ->
       val igEntity = igDao.getImplementationGuide(igDependency.packageId, igDependency.version)
-      igDao.deleteImplementationGuide(igEntity)
-      igEntity.rootDirectory.deleteRecursively()
+      if (igEntity != null) {
+        igDao.deleteImplementationGuide(igEntity)
+        igEntity.rootDirectory.deleteRecursively()
+      }
     }
   }
+
+  private suspend fun importFile(igId: Long, file: File) {
+    val resource = withContext(Dispatchers.IO) {
+      try {
+        FileInputStream(file).use(jsonParser::parseResource)
+      } catch (exception: Exception) {
+        Timber.d(exception, "Unable to import file: %file")
+      }
+    }
+    when (resource) {
+      is Bundle -> importBundle(igId, resource, file)
+      is Resource -> importResource(igId, resource, file)
+    }
+  }
+
 
   private suspend fun importBundle(igId: Long, resource: Bundle, file: File) {
     // TODO: Multiple resources will point to the same file, and it needs to be accounted for while
@@ -127,5 +141,24 @@ class IgManager internal constructor(igDatabase: ImplementationGuideDatabase) {
 
   private fun loadResource(resourceEntity: ResourceMetadataEntity): IBaseResource {
     return jsonParser.parseResource(FileInputStream(resourceEntity.resourceFile))
+  }
+
+  fun close() {
+    igDatabase.close()
+  }
+
+  companion object {
+    val DEFAULT_DEPENDENCY =
+      IgDependency("com.google.android.fhir", "1.0.0", "http://github.com/google/android-fhir")
+
+    fun create(context: Context) = IgManager(Room.databaseBuilder(context,
+                                                                  ImplementationGuideDatabase::class.java,
+                                                                  "implementationguide.db")
+                                               .build())
+
+    fun createInMemory(context: Context) =
+      IgManager(Room.inMemoryDatabaseBuilder(context, ImplementationGuideDatabase::class.java)
+                  .build())
+
   }
 }
